# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Locação API",
    "version": "1.0.0",
    "description": "API para gerenciamento de locações e brinquedos."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor Local"
    }
  ],
  "paths": {
...
    "/tiposBrinquedos": {
      "get": {
        "summary": "Listar todos os tipos de brinquedos",
        "operationId": "getTiposBrinquedos",
        "responses": {
          "200": {
            "description": "Lista de tipos de brinquedos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TipoBrinquedoResponseDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar um novo tipo de brinquedo",
        "operationId": "createTipoBrinquedo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTipoBrinquedoDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tipo de brinquedo criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoBrinquedoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/tiposBrinquedos/{cod}": {
      "parameters": [
        {
          "name": "cod",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          }
        }
      ],
      "get": {
        "summary": "Obter um tipo de brinquedo pelo código",
        "operationId": "getTipoBrinquedoById",
        "responses": {
          "200": {
            "description": "Tipo de brinquedo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoBrinquedoResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar um tipo de brinquedo",
        "operationId": "updateTipoBrinquedo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTipoBrinquedoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tipo de brinquedo atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoBrinquedoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTipoBrinquedoDTO": {
        "type": "object",
        "required": ["nome"],
        "properties": {
          "nome": {
            "type": "string",
            "example": "Carrinho"
          }
        },
        "additionalProperties": false
      },
      "UpdateTipoBrinquedoDTO": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TipoBrinquedoResponseDTO": {
        "type": "object",
        "properties": {
          "cod": {
            "type": "string",
            "format": "uuid",
            "example": "8e07cb00-089a-4e0f-b366-3889bf2c61db"
          },
          "nome": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}